Parse command line args for elmy binary
Development of the elmy binary has been started with the ability to
process the command line arguments in both long and short forms. The
handlers for each argument have also been prepared, but are incomplete
as of now except for the handlers for -h and -v. The remaining handlers
will be developed as the library interface is implemented.

Signed-off-by: Abhishek Chakravarti <abhishek@taranjali.org>