Complete binary implementation
The `elmy` binary is now complete in its implementation, barring a few
minor details that will be fixed in the next commits. The binary code
base has been overhauled to be easier to understand and maintain. Quick
sanity checks have been performed manually, but formal testing is
pending as yet.

Signed-off-by: Abhishek Chakravarti <abhishek@taranjali.org>